+++ Cargo.toml
@@ -38,6 +38,13 @@ indexmap = "2.7.1"
 time = "0.3.37"
 tower-sessions = { version = "0.14.0" }
 url = "2.5"
+tower-http = { version = "0.6.1", features = [
+  "catch-panic",
+  "timeout",
+  "sensitive-headers",
+  "cors",
+  "trace",
+] }
 tokio = { version = "1", features = ["full"] }
 async-trait = { version = "0.1.84" }
 serde = { version = "1.0", features = ["derive"] }
