+++ crates/api-ui/src/tests/schemas.rs
@@ -66,8 +66,7 @@ async fn test_ui_schemas() {
     )
     .await
     .unwrap();
-    // assert_eq!(http::StatusCode::OK, res.clone().status());
-    println!("res: {:#?}", res.bytes().await.unwrap());
+    assert_eq!(http::StatusCode::OK, res.status());
 
     let schema_name = "testing2".to_string();
     let payload2 = SchemaCreatePayload {
